TIP-856: Move Pim\ElasticsearchBundle spec to Akeneo\Tool
