+++ src/Akeneo/Tool/Bundle/ElasticsearchBundle/spec/IndexConfiguration/conf1.yml
