+++ src/Akeneo/Tool/Bundle/ElasticsearchBundle/spec/IndexConfiguration/conf2.yml
