+++ src/Akeneo/Tool/Bundle/ElasticsearchBundle/spec/IndexConfiguration/conf3.yml
